+++ sample/src/main/java/mobi/parchment/MenuActivity.java
@@ -35,5 +35,4 @@ public class MenuActivity extends Activity {
         startActivity(SimpleViewPagerActivity.class);
     }
 
-
 }