+++ sample/src/main/java/mobi/parchment/SimpleApplication.java
@@ -1,7 +1,0 @@
-package mobi.parchment;
-
-import android.app.Application;
-
-public class SimpleApplication extends Application {
-
-}