+++ sample/src/main/java/mobi/parchment/SimpleGridPatternViewActivity.java
@@ -1,7 +1,6 @@
 package mobi.parchment;
 
 import android.os.Bundle;
-import android.widget.AdapterView;
 import android.widget.BaseAdapter;
 
 import mobi.parchment.widget.adapterview.gridpatternview.GridPatternView;
@@ -12,20 +11,18 @@ import java.util.List;
 
 public class SimpleGridPatternViewActivity extends BaseActivity {
 
-    private GridPatternView<BaseAdapter> mGridPatternView;
-
     @SuppressWarnings("unchecked")
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_simple_grid_pattern_view);
 
-        mGridPatternView = (GridPatternView<BaseAdapter>) findViewById(R.id.parchment_view);
+        final GridPatternView<BaseAdapter> gridPatternView = (GridPatternView<BaseAdapter>) findViewById(R.id.parchment_view);
 
         final List<GridPatternItemDefinition> gridPatternItemDefinitions = new ArrayList<GridPatternItemDefinition>();
 
         gridPatternItemDefinitions.add(new GridPatternItemDefinition(0, 0, 6, 4));
-        mGridPatternView.addGridPatternGroupDefinition(gridPatternItemDefinitions);
+        gridPatternView.addGridPatternGroupDefinition(gridPatternItemDefinitions);
         gridPatternItemDefinitions.clear();
 
         gridPatternItemDefinitions.add(new GridPatternItemDefinition(0, 0, 2, 2));
@@ -33,7 +30,7 @@ public class SimpleGridPatternViewActivity extends BaseActivity {
         gridPatternItemDefinitions.add(new GridPatternItemDefinition(4, 0, 2, 3));
         gridPatternItemDefinitions.add(new GridPatternItemDefinition(0, 2, 4, 4));
         gridPatternItemDefinitions.add(new GridPatternItemDefinition(4, 3, 2, 3));
-        mGridPatternView.addGridPatternGroupDefinition(gridPatternItemDefinitions);
+        gridPatternView.addGridPatternGroupDefinition(gridPatternItemDefinitions);
         gridPatternItemDefinitions.clear();
 
         gridPatternItemDefinitions.add(new GridPatternItemDefinition(0, 0, 2, 3));
@@ -41,13 +38,13 @@ public class SimpleGridPatternViewActivity extends BaseActivity {
         gridPatternItemDefinitions.add(new GridPatternItemDefinition(0, 3, 2, 3));
         gridPatternItemDefinitions.add(new GridPatternItemDefinition(2, 4, 2, 2));
         gridPatternItemDefinitions.add(new GridPatternItemDefinition(4, 4, 2, 2));
-        mGridPatternView.addGridPatternGroupDefinition(gridPatternItemDefinitions);
+        gridPatternView.addGridPatternGroupDefinition(gridPatternItemDefinitions);
 
-        mGridPatternView.setAdapter(getProductsAdapter());
+        gridPatternView.setAdapter(getProductsAdapter());
     }
 
     @Override
-    public AdapterView<?> getAdapterView() {
-        return mGridPatternView;
+    public int getLayoutResourceId() {
+        return R.layout.list_item_gridpatternview_picture;
     }
 }