+++ sample/src/main/java/mobi/parchment/SimpleViewPagerActivity.java
@@ -1,14 +1,12 @@
 package mobi.parchment;
 
 import android.os.Bundle;
-import android.widget.AdapterView;
 import android.widget.BaseAdapter;
 
 import mobi.parchment.widget.adapterview.listview.ListView;
 
 public class SimpleViewPagerActivity extends BaseActivity{
 
-	private ListView<BaseAdapter> mViewPager;
 
 	@SuppressWarnings("unchecked")
 	@Override
@@ -16,13 +14,13 @@ public class SimpleViewPagerActivity extends BaseActivity{
 		super.onCreate(savedInstanceState);
 		setContentView(R.layout.activity_simple_view_pager);
 
-		mViewPager = (ListView<BaseAdapter>) findViewById(R.id.parchment_view);
-		mViewPager.setAdapter(getProductsAdapter());
+		final ListView viewPager = (ListView<BaseAdapter>) findViewById(R.id.parchment_view);
+		viewPager.setAdapter(getProductsAdapter());
 	}
 
     @Override
-    public AdapterView<?> getAdapterView() {
-        return mViewPager;
+    public int getLayoutResourceId() {
+        return R.layout.list_item_view_pager_picture;
     }
 
 }