+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 #Change Log
 
+##Version 1.6.6
+
+  * Revamped Sample application
+  * Fixed bug with snapToPosition and start
+  * Fixed bug with ViewPager constantly paging when a view is larger than the screen
+
 ##Version 1.6.5
 
   * Removed CellPadding from outter portion, this allows users to control how close their cells are to the sides of the view as well as spacing between the views themselves.